Restructuracion segun consignas Cristian

+++ src/controller/products.controller.ts
@@ -20,7 +20,8 @@ export class ProductsController {
         }
         catch(error) {
             errorLog.error(error)
-            res.json({error: "There has been an error fetching the products."});
+            res.status(500);
+            res.json({error: `There has been an error fetching the products => ${error}`});
         }
     }
 
@@ -33,7 +34,8 @@ export class ProductsController {
         }
         catch(error) {
             errorLog.error(error)
-            res.json({error: "There has been an error fetching the products."})
+            res.status(500);
+            res.json({error: `There has been an error fetching the products => ${error}`})
         }
     }
 
@@ -46,50 +48,87 @@ export class ProductsController {
         }
         catch (error){
             errorLog.error(error);
+            res.status(500);
             res.json({error: "There has been an error fetching the product."})
         }
     }
 
-    addProduct = async (req: Request, res: Response) => {
+    addProduct = async (req: any, res: Response) => {
         try {
-            const {name, category, description, foto, price} = req.body;
-            let product = new Product(name, category, description, foto, price)
-            const resultValidation = await validate(product);
-            if(resultValidation.length > 0){
-                errorLog.error(resultValidation);
-                res.json({});
+            const user = req.user;
+            if(user.admin === false){
+                res.status(401);
+                res.json({error: `You must be an admin user to use this route`})
             } else {
-                let addedProduct = await this.apiProducts.addProduct(product);
-                res.json(addedProduct);
+                const {name, category, description, fotos, price, stock} = req.body;
+                let product = new Product(name, category, description, fotos, price, stock)
+                const resultValidation = await validate(product);
+                if(resultValidation.length > 0){
+                    errorLog.error(resultValidation);
+                    res.status(400);
+                    res.json({error: `You must provide the required data => ${resultValidation}`});
+                } else {
+                    let addedProduct = await this.apiProducts.addProduct(product);
+                    res.status(201);
+                    res.json(addedProduct);
+                }
             }
         }
         catch (error){
             errorLog.error(error);
+            res.status(500);
             res.json({error: "There has been an error saving the product."})
         }
     }
 
-    updateProductById = async(req: Request, res: Response) => {
+    updateProductById = async(req: any, res: Response) => {
         try {
-            const {id} = req.params;
-            const {name, category, description, foto, price} = req.body;
-            let product = new Product(name, category, description, foto, price)
-            let modifiedProduct = await this.apiProducts.updateProductById(id, product);
-            res.json(modifiedProduct);
+            const user = req.user;
+            if(user.admin === false){
+                res.status(401);
+                res.json({error: `You must be an admin user to use this route`})
+            } else{
+                const {id} = req.params;
+                const {name, category, description, fotos, price, stock} = req.body;
+                let product = new Product(name, category, description, fotos, price, stock)
+                const resultValidation = await validate(product);
+                if(resultValidation.length > 0){
+                    errorLog.error(resultValidation);
+                    res.status(400);
+                    res.json({error: `You must provide the required data => ${resultValidation}`});
+                } else {
+                    let modifiedProduct = await this.apiProducts.updateProductById(id, product);
+                    res.status(200);
+                    res.json(modifiedProduct);
+                }
+            }
         } catch (error){
             errorLog.error(error);
             res.json({error: "There has been an error updating the product."})
         }
     }
 
-    deleteProduct = async(req: Request, res: Response) => {
+    deleteProduct = async(req: any, res: Response) => {
         try{
-            const {id} = req.params;
-            let deletedProduct = await this.apiProducts.deleteProduct(id);
-            res.json(deletedProduct);
+            const user = req.user;
+            if(user.admin === false){
+                res.status(401);
+                res.json({error: `You must be an admin user to use this route`})
+            } else{
+                const {id} = req.params;
+                let deletedProduct = await this.apiProducts.deleteProduct(id);
+                if(deletedProduct == {}){
+                    res.status(404);
+                    res.json({})
+                } else {
+                    res.status(200)
+                    res.json(deletedProduct);
+                }
+            }
         }
         catch (error){
             errorLog.error(error)
+            res.status(500);
             res.json({error: "There has been an error deleting the product."})
         }
     }
